+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added New Function to get the IDs of the Users.
 - Fixed Users updating all instead of selected ones.
 - Added DELETE endpoint for deleting Single User.
+- Added Switcher Button to Admin only.
 
 ## 31-07-2024
 - Fixed True/False status in entire instance of the App.